pyproject.toml
[project]
name = "car-rent-db"
version = "0.1.0"
description = ""
authors = [
    {name = "s-sorokaty",email = "sorokaty0824@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "alembic (>=1.16.2,<2.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "pytest (>=8.4.1,<9.0.0)",
    "psycopg[binary] (>=3.2.9,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
